+++ rtiow/rttnw/sphere.h
@@ -310,8 +310,8 @@ class rotate_y : public hittable {
 
 rotate_y::rotate_y(hittable *p, float angle) : ptr(p) {
     float radians = (M_PI / 180.) * angle;
-    sin_theta = sin(radians);
-    cos_theta = cos(radians);
+    sin_theta = sinf(radians);
+    cos_theta = cosf(radians);
     hasbox = ptr->bounding_box(0, 1, bbox);
     vec3 min(FLT_MAX, FLT_MAX, FLT_MAX);
     vec3 max(-FLT_MAX, -FLT_MAX, -FLT_MAX);
